move toJsonSchema to JsonSchema file
