+++ internal/plugin/providers/openapi/source.go
@@ -9,7 +9,6 @@ import (
 	"fmt"
 	"os"
 	"path/filepath"
-	"slices"
 	"strings"
 
 	"github.com/marmotdata/marmot/internal/core/asset"
@@ -18,9 +17,7 @@ import (
 	"github.com/marmotdata/marmot/internal/plugin"
 	"github.com/pb33f/libopenapi"
 	"github.com/pb33f/libopenapi/datamodel"
-	"github.com/pb33f/libopenapi/datamodel/high/base"
 	v3 "github.com/pb33f/libopenapi/datamodel/high/v3"
-	"github.com/pb33f/libopenapi/renderer"
 	"github.com/rs/zerolog/log"
 )
 
@@ -276,7 +273,7 @@ func (s *Source) createEndpointAssets(spec*libopenapi.DocumentModel[v3.Document]
 			schema := make(map[string]string)
 			for code, response := range op.Responses.Codes.FromOldest() {
 				for content, mediaType := range response.Content.FromOldest() {
-					jsonSchema, err := toJsonSchema(mediaType.Schema)
+					jsonSchema, err := NewJsonSchemaFromOpenAPISchema(mediaType.Schema)
 					if err != nil {
 						log.Warn().Err(err).Msg("Failed to convert OpenAPI schema to json schema")
 						continue
@@ -346,161 +343,4 @@ func isYAML(path string) bool {
 	return ext == ".yaml" || ext == ".yml"
 }
 
-func toJsonSchema(schemaProxy *base.SchemaProxy) (*JsonSchema, error) {
-	root := NewRootJsonSchema()
-	mg := mockGenerator()
 
-	var dfs func(p *base.SchemaProxy, r *JsonSchema, depth int) (*JsonSchema, error)
-	dfs = func(p *base.SchemaProxy, r *JsonSchema, depth int) (*JsonSchema, error) {
-		jSchema := NewJsonSchema()
-
-		if p == nil {
-			return &jSchema, nil
-		}
-
-		if depth > 5 {
-			description := "Maximum depth reached. Stop rendering schema"
-			jSchema.Description = &description
-			return &jSchema, nil
-		}
-
-		schema, err := p.BuildSchema()
-		if err != nil {
-			return nil, fmt.Errorf("failed to build schema at line %d: %w", p.GetValueNode().Line, err)
-		}
-		if len(schema.Type) > 0 {
-			jSchema.Type = schema.Type
-		}
-		jSchema.Description = &schema.Description
-
-		if schema.Nullable != nil && *schema.Nullable {
-			if jSchema.Type != nil {
-				jSchema.Type = append(jSchema.Type, "null")
-			} else {
-				jSchema.Type = []string{"null"}
-			}
-		}
-
-		if schema.Properties != nil {
-			jSchema.Properties = make(map[string]JsonSchema)
-			for name, prop := range schema.Properties.FromOldest() {
-				jSubSchema, err := dfs(prop, r, depth + 1)
-				if err != nil {
-					return nil, fmt.Errorf("failed to render properties schema at line %d: %w", prop.GetValueNode().Line, err)
-				}
-				jSchema.Properties[name] = *jSubSchema
-			}
-		}
-
-		if schema.Items != nil && slices.Contains(schema.Type, "array") {
-			isSchemaProxy := schema.Items.IsA()
-			if isSchemaProxy {
-				jsonItemsSchema, err := dfs(schema.Items.A, r, depth + 1)
-				if err != nil {
-					return nil, fmt.Errorf("failed to render items schema at line %d: %w", schema.Items.A.GetValueNode().Line, err)
-				}
-				jSchema.Items = jsonItemsSchema
-			} else {
-				jSchema.Items = schema.Items.B
-			}
-		}
-
-		if schema.AllOf != nil {
-			jSchema.AllOf = []*JsonSchema{}
-			for _, allOfSchema := range schema.AllOf {
-				jsonAllOfSchema, err := dfs(allOfSchema, r, depth + 1)
-				if err != nil {
-					return nil, fmt.Errorf("failed to render AllOf schema at line %d: %w", allOfSchema.GetValueNode().Line, err)
-				}
-				jSchema.AllOf = append(jSchema.AllOf, jsonAllOfSchema)
-			}
-		}
-
-		if schema.AnyOf != nil {
-			jSchema.AnyOf = []*JsonSchema{}
-			for _, anyOfSchema := range schema.AnyOf {
-				jsonAnyOfSchema, err := dfs(anyOfSchema, r, depth + 1)
-				if err != nil {
-					return nil, fmt.Errorf("failed to render AnyOf schema at line %d: %w", anyOfSchema.GetValueNode().Line, err)
-				}
-				jSchema.AnyOf = append(jSchema.AnyOf, jsonAnyOfSchema)
-			}
-		}
-
-		if schema.OneOf != nil {
-			jSchema.OneOf = []*JsonSchema{}
-			for _, oneOfSchema := range schema.OneOf {
-				jsonOneOfSchema, err := dfs(oneOfSchema, r, depth + 1)
-				if err != nil {
-					return nil, fmt.Errorf("failed to render OneOf schema at line %d: %w", oneOfSchema.GetValueNode().Line, err)
-				}
-				jSchema.OneOf = append(jSchema.OneOf, jsonOneOfSchema)
-			}
-		}
-
-		if schema.Required != nil {
-			jSchema.Required = schema.Required
-		}
-
-		if schema.Enum != nil {
-			enum := []any{}
-			for _, e := range schema.Enum {
-				enum = append(enum, e.Value)
-			}
-			jSchema.Enum = enum
-		}
-
-		if schema.AdditionalProperties != nil {
-			if schema.AdditionalProperties.IsA() {
-				jSchema.AdditionalProperties = schema.AdditionalProperties.A
-			} else {
-				jSchema.AdditionalProperties = schema.AdditionalProperties.B
-			}
-		}
-
-		if schema.Not != nil {
-			jsonNotSchema, err := dfs(schema.Not, r, depth + 1)
-			if err != nil {
-				return nil, fmt.Errorf("failed to render Not schema at line %d: %w", schema.Not.GetValueNode().Line, err)
-			}
-			jSchema.Not = jsonNotSchema
-		}
-
-		jSchema.Pattern = schema.Pattern
-		jSchema.MultipleOf = schema.MultipleOf
-		jSchema.Maximum = schema.Maximum
-		jSchema.Minimum = schema.Minimum
-		jSchema.MaxLength = schema.MaxLength
-		jSchema.MinLength = schema.MinLength
-		jSchema.MaxItems = schema.MaxItems
-		jSchema.MinItems = schema.MinItems
-		jSchema.UniqueItems = schema.UniqueItems
-		jSchema.MaxProperties = schema.MaxProperties
-		jSchema.MinProperties = schema.MinProperties
-		jSchema.MinProperties = schema.MinProperties
-		jSchema.Deprecated = schema.Deprecated
-
-		mock, err := mg.GenerateMock(schema, "")
-		if err != nil {
-			if err.Error() != "unable to render schema for mock, it's empty" {
-				log.Warn().Err(err).Msg(fmt.Sprintf("Failed to generate mock for schema at line %d", p.GetValueNode().Line))
-			} 
-		}
-		jSchema.Example = string(mock)
-
-		return &jSchema, nil
-	}
-
-	jsonSchema, err := dfs(schemaProxy, &root, 0)
-	if err != nil {
-		return nil, fmt.Errorf("Failed to dive into schema at line %d: %w", schemaProxy.GetValueNode().Line, err)
-	}
-
-	return jsonSchema, nil 
-}
-
-func mockGenerator() *renderer.MockGenerator {
-	mg := renderer.NewMockGenerator(renderer.JSON)
-	mg.SetPretty()
-	return mg
-}
